#310: Changed ByteStream.isAvailable(...)'s length argument to BigInteger to allow Defs of arbitrary size.

+++ core/src/main/java/io/parsingdata/metal/data/ByteStream.java
@@ -23,6 +23,6 @@ public interface ByteStream {
 
     byte[] read(BigInteger offset, int length) throws IOException;
 
-    boolean isAvailable(BigInteger offset, int length);
+    boolean isAvailable(BigInteger offset, BigInteger length);
 
 }
